requirements.txt
Django>=4.2
python-dotenv
pandas
requests
openpyxl
pymysql

